fix: correctly configure redis instance
Superset was ignoring the cairn-redis container, so we decided to remove it.
But the superset celery worker was writing to the /0 database, which conflicts
with the default celery database from the lms/cms. We fix this issue by running
in a different database.
The CELERY_RESULT_BACKEND seemed to be completely useless, so we got rid of it.